Fix correctness bugs from Fable review (#72, #73, #74, #75)
#72 store_url never persisted: fetch_and_store now copies metadata.store_url
onto the record, so enrolled extensions get a real store URL instead of the
empty placeholder — fixes the broken "View in store" link and webhook payloads.

#73 /login ⇄ / redirect loop: login_page only did a signature/age cookie check
while require_auth does a DB-backed check, so a signature-valid but stale cookie
(deleted user or post-password-change) looped forever. Factor the shared
authenticate_session() helper used by both require_auth and require_api_auth,
have login_page use it (redirect home only for a valid user), and clear the
stale cookie so the loop can't re-form.

#74 unreachable obfuscation tier: every single-char identifier is also a <=2-char
one, so testing short first made the single-char branch dead. Check the stronger
single-char signal first (+4), then the <=2-char signal (+3).

#75 orphaned placeholder: _enroll_extension only cleaned up on FetchError; any
other first-fetch failure left an unscored placeholder on the watchlist and 500'd.
Add an except-Exception arm (rollback → discard placeholder → re-raise) via the
shared _discard_placeholder helper.

Regression tests added for each. Full suite: 303 passed.

Fixes #72
Fixes #73
Fixes #74
Fixes #75

+++ app/inspector.py
@@ -507,9 +507,13 @@ def _obfuscation_score(source: str) -> int:
     single = sum(1 for i in identifiers if len(i) == 1)
 
     if total > 50:
-        if short / total > 0.6:
+        # Check the stronger single-char signal first: every single-char identifier
+        # is also a <=2-char (short) one, so `short >= single` always. Testing short
+        # first made the single-char branch unreachable (#74); order them so the
+        # heavier obfuscation (mostly one-letter names) scores highest.
+        if single / total > 0.6:
             score += 4
-        elif single / total > 0.6:
+        elif short / total > 0.6:
             score += 3
 
     # High ratio of escaped unicode or hex sequences
